test: Consolidate modules for testing parameter dicts
